Another bug change check in,  this is a fix for translations, added and removed  fields on some of the models,  new database  migration scripts, stoped the casting of Other Costs across all view models that was causing issues.

+++ PigTool/PigTool/ViewModels/DataViewModels/HealthCareViewModel.cs
@@ -546,7 +546,7 @@ namespace PigTool.ViewModels.DataViewModels
                 _itemForEditing.OtherPurchasedFrom = OtherPurchasedFrom;
                 //_itemForEditing.CreatedBy = User.AuthorisedUserName;
                 _itemForEditing.TransportationCost = (double)TransportationCost;
-                _itemForEditing.OtherCosts = (double)OtherCosts;
+                _itemForEditing.OtherCosts = OtherCosts;
                 _itemForEditing.Comment = Comment;
                 _itemForEditing.LastModified = DateTime.UtcNow;
