+++ PigTool/PigTool/ViewModels/DataViewModels/ManureSaleViewModel.cs
@@ -422,7 +422,7 @@ namespace PigTool.ViewModels.DataViewModels
                 _itemForEditing.PaymentType = SelectedPaymentType != null ? SelectedPaymentType.TranslationRowKey : null;
                 _itemForEditing.PaymentValue = (double)PaymentValue;
                 _itemForEditing.TransportationCost = (double)TransportationCost;
-                _itemForEditing.OtherCosts = (double)OtherCosts;
+                _itemForEditing.OtherCosts = OtherCosts;
                 _itemForEditing.Comment = Comment;
                 _itemForEditing.LastModified = DateTime.UtcNow;
 
@@ -443,7 +443,7 @@ namespace PigTool.ViewModels.DataViewModels
                     PaymentType = SelectedPaymentType != null ? SelectedPaymentType.TranslationRowKey : null,
                     PaymentValue = PaymentValue,
                     TransportationCost = (double)TransportationCost,
-                    OtherCosts = (double)OtherCosts,
+                    OtherCosts = OtherCosts,
                     Comment = Comment,
                     LastModified = DateTime.UtcNow,
                     CreatedBy = User.UserName,
